Making the component compatible with react native >= 0.40

+++ android/src/main/java/com/joshblour/reactnativeheading/ReactNativeHeadingModule.java
@@ -83,8 +83,10 @@ public class ReactNativeHeadingModule extends ReactContextBaseJavaModule impleme
             // calculate th rotation matrix
             SensorManager.getRotationMatrixFromVector(rMat, event.values);
             // get the azimuth value (orientation[0]) in degree
-            newAzimuth = (((( Math.toDegrees( SensorManager.getOrientation( rMat, orientation )[0] ) + 360 ) % 360) -
-                          ( Math.toDegrees( SensorManager.getOrientation( rMat, orientation )[2] ))) +360) % 360;
+
+            newAzimuth = (int) ((((( Math.toDegrees( SensorManager.getOrientation( rMat, orientation )[0] ) + 360 ) % 360) -
+                          ( Math.toDegrees( SensorManager.getOrientation( rMat, orientation )[2] ))) +360) % 360);
+
             //dont react to changes smaller than the filter value
             if (Math.abs(mAzimuth - newAzimuth) < mFilter) {
                 return;
@@ -95,9 +97,6 @@ public class ReactNativeHeadingModule extends ReactContextBaseJavaModule impleme
                     .emit("headingUpdated", (int) newAzimuth);
 
             mAzimuth = newAzimuth;
-            // Log.e("Azimuth",
-            //   String.valueOf(newAzimuth));
-            //
         }
     }
